Moved functions that do API requests to a file called api_client so that utils doesnt get bloated
